Registers Animation Logic Error 01
Previously, the program was using just 1 text block to display information being animated between the registers and the processor. The error encountered was the problem of how to deal with more than 1 pieces of data travelling from different registers simultaneously.
Solution:
Now there is an array of text blocks to be used, one for each register on the chip.

Work has begun, not yet complete